requirements.txt
appnope==0.1.0
arrow==0.10.0
bleach==1.5.0
dateutils==0.6.6
decorator==4.0.11
Django==1.10.5
django-extensions==1.7.6
djangorestframework==3.5.3
entrypoints==0.2.2
html5lib==0.9999999
ipykernel==4.5.2
ipython==5.2.2
ipython-genutils==0.1.0
ipywidgets==5.2.2
Jinja2==2.9.5
jsonschema==2.5.1
jupyter-client==4.4.0
jupyter-core==4.2.1
MarkupSafe==0.23
mistune==0.7.3
nbconvert==5.1.1
nbformat==4.2.0
notebook==4.3.2
numpy==1.12.0
pandas==0.19.2
pandocfilters==1.4.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.13
ptyprocess==0.5.1
Pygments==2.2.0
python-dateutil==2.6.0
pytz==2016.10
pyzmq==16.0.2
simplegeneric==0.8.1
six==1.10.0
terminado==0.6
testpath==0.3
tornado==4.4.2
traitlets==4.3.1
wcwidth==0.1.7
widgetsnbextension==1.2.6
